% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular.R
\name{print.angular}
\alias{print.angular}
\alias{coef.angular}
\alias{vcov.angular}
\alias{fitted.angular}
\alias{residuals.angular}
\alias{predict.angular}
\alias{summary.angular}
\alias{print.summary.angular}
\alias{plot.angular}
\title{Methods for Angular Regression Objects}
\usage{
\method{print}{angular}(x, ...)

\method{coef}{angular}(object, ...)

\method{vcov}{angular}(object, robust = FALSE, ...)

\method{fitted}{angular}(object, ...)

\method{residuals}{angular}(object, ...)

\method{predict}{angular}(object, newdata = NULL, type = c("response", "components"), ...)

\method{summary}{angular}(object, ...)

\method{print}{summary.angular}(x, ...)

\method{plot}{angular}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"angular"}.}

\item{...}{Additional arguments.}

\item{object}{An object of class \code{"angular"}.}

\item{robust}{Logical; for \code{vcov()}, return the sandwich covariance
matrix when \code{TRUE}.}

\item{newdata}{Optional data frame for prediction.}

\item{type}{Prediction type. Use \code{"response"} for fitted angles in
radians or \code{"components"} for cosine and sine components.}
}
\description{
Methods for Angular Regression Objects
}
