% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular_two_step.R
\name{print.angular_two_step}
\alias{print.angular_two_step}
\alias{summary.angular_two_step}
\alias{print.summary.angular_two_step}
\alias{coef.angular_two_step}
\alias{fitted.angular_two_step}
\alias{residuals.angular_two_step}
\alias{predict.angular_two_step}
\alias{plot.angular_two_step}
\alias{logLik.angular_two_step}
\alias{AIC.angular_two_step}
\alias{BIC.angular_two_step}
\title{Methods for two-step angular regression objects}
\usage{
\method{print}{angular_two_step}(x, ...)

\method{summary}{angular_two_step}(object, ...)

\method{print}{summary.angular_two_step}(x, ...)

\method{coef}{angular_two_step}(object, step = c("homogeneous", "consensus"), ...)

\method{fitted}{angular_two_step}(object, ...)

\method{residuals}{angular_two_step}(object, ...)

\method{predict}{angular_two_step}(object, newdata = NULL, type = c("response", "components"), ...)

\method{plot}{angular_two_step}(x, ...)

\method{logLik}{angular_two_step}(object, ...)

\method{AIC}{angular_two_step}(object, ..., k = 2)

\method{BIC}{angular_two_step}(object, ...)
}
\arguments{
\item{x, object}{An object of class \code{"angular_two_step"}.}

\item{...}{Additional arguments passed to the underlying method.}

\item{step}{Which fitted step to use for \code{coef()}.}

\item{newdata}{Optional data frame for prediction.}

\item{type}{Prediction type passed to \code{\link{predict.angular}()}.}

\item{k}{Numeric penalty used by \code{AIC()}.}
}
\description{
Methods for two-step angular regression objects
}
