% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoregressivedata.R
\name{autoregressivedata}
\alias{autoregressivedata}
\title{Create autoregressive dataset}
\usage{
autoregressivedata(data, response, p, ID = NULL)
}
\arguments{
\item{data}{A data frame containing at least the response variable.}

\item{response}{A character string specifying the name of the angular response variable.}

\item{p}{Integer giving the order of the autoregressive structure.}

\item{ID}{Optional. A character string specifying the name of a grouping variable. If provided, lags are computed within each group.}
}
\value{
The input \code{data} with \code{p} additional columns named \code{response_tmk} for \eqn{k=1,\ldots,p} containing the lagged values of \code{response}.
}
\description{
This function augments a data frame with lagged versions of an angular response
variable so that autoregressive circular regression models can be fitted.
}
