% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircularRegression-package.R
\docType{package}
\name{CircularRegression-package}
\alias{CircularRegression}
\alias{CircularRegression-package}
\title{CircularRegression: Regression models for circular data}
\description{
This package implements various regression models for circular data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AurelienNicosiaULaval/CircularRegression}
  \item Report bugs at \url{https://github.com/AurelienNicosiaULaval/CircularRegression/issues}
}

}
\author{
\strong{Maintainer}: Aurelien Nicosia \email{aurelien.nicosia@mat.ulaval.ca}

}
\keyword{internal}
