% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_reference_angle.r
\name{.extract_angles_from_formula}
\alias{.extract_angles_from_formula}
\title{Extract unique angle names (left of "x:z"), preserving first appearance order}
\usage{
.extract_angles_from_formula(formula)
}
\description{
Example:
  y ~ x.meadow + x.meadow:z.meadow + x.gap + x.gap:z.gap
-> c("x.meadow","x.gap")
}
