% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_reference_angle.r
\name{print.pick_reference_angle}
\alias{print.pick_reference_angle}
\title{Print method for pick_reference_angle objects}
\usage{
\method{print}{pick_reference_angle}(x, digits = 4, ...)
}
\arguments{
\item{x}{A list returned by pick_reference_angle()}

\item{digits}{Number of digits to print (default 4)}

\item{...}{Unused}
}
\description{
Print method for pick_reference_angle objects
}
