% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sandhopper.r
\docType{data}
\name{Sandhopper}
\alias{Sandhopper}
\title{Sandhopper Escape Orientation Data}
\format{

}
\usage{
Sandhopper
}
\description{
This dataset contains repeated measurements on the escape orientation
of 59 sandhoppers (*Talitrus saltator*) in a circular arena experiment
with a diameter of 120 cm. Each individual was released five times, and
its escape angle with respect to the north was recorded. Several
explanatory variables were also measured during the experiment.
}
\details{
The dataset was originally presented and analyzed by D’Elia (2001) and
later used in Rivest & Kato (2016) to illustrate angular regression
models with repeated measures.
}
\references{
D’Elia, G. (2001). Orientation in sandhoppers. *Behavioural Ecology*.
Rivest, L.-P., & Kato, S. (2016). A random effects model for clustered circular data.
  *The Canadian Journal of Statistics*.
}
\keyword{datasets}
