% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular.R
\name{print.angular}
\alias{print.angular}
\alias{coef.angular}
\alias{residuals.angular}
\alias{summary.angular}
\alias{print.summary.angular}
\alias{plot.angular}
\title{Methods for Angular Regression Objects}
\usage{
\method{print}{angular}(x, ...)

\method{coef}{angular}(object, ...)

\method{residuals}{angular}(object, ...)

\method{summary}{angular}(object, ...)

\method{print}{summary.angular}(x, ...)

\method{plot}{angular}(x, ...)
}
\arguments{
\item{x}{An object of class "angular".}

\item{...}{Further arguments passed to plotting functions.}

\item{object}{An object of class "angular".}
}
\value{
An object of class "summary.angular".
}
\description{
These functions provide methods for printing, extracting coefficients,
summarizing, and retrieving residuals from an angular regression model.

Summarizes an object of class "angular" by providing key diagnostic statistics.

This function produces diagnostic plots for an object of class "angular".
}
