% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular.R
\name{angular}
\alias{angular}
\title{Angular Regression Model}
\usage{
angular(formula, data, model = "simplified", initbeta = NULL, control = list())
}
\arguments{
\item{formula}{A formula with the dependent angle on the left of the ~ operator and terms specifying
the explanatory variables on the right. These terms must be written \code{x:z}, where
\code{x} is an explanatory angle whose relative importance might depend on the
positive variable \code{z}. For \code{model = "simplified"}, the first explanatory angle is
the reference direction (any provided \code{z} for this angle is ignored).}

\item{data}{An optional data frame, list or environment containing the variables in the model formula.}

\item{model}{A character string, either \code{"complete"} for the complete model with an intercept (the default)
or \code{"simplified"} for the simplified model without an intercept.}

\item{initbeta}{A numerical vector of initial beta parameter values. The default is to use the best initial
values found among a grid of possible values.}

\item{control}{A list of control parameters. See Details.}
}
\value{
An object of class "angular" containing:
\describe{
  \item{MaxCosine}{the maximum cosine value.}
  \item{parameters}{the parameter estimates and their standard errors, z-values and associated p-values.}
  \item{varcov0}{the asymptotic estimated variance-covariance matrix.}
  \item{varcov1}{the non-parametric estimated variance-covariance matrix (sandwich).}
  \item{long}{the vector of predicted concentrations.}
  \item{mui}{the vector of predicted mean angles.}
  \item{y}{the response variable.}
  \item{iter.detail}{the iteration details.}
  \item{call}{the function call.}
}
}
\description{
Function that fits a regression model for angular variables along with its
associated S3 methods.
}
\details{
The control argument is a list that can supply any of the following components:
\describe{
  \item{\code{pginit}}{The approximate number of points on the grid of possible initial beta values tried when
                        \code{initbeta} is not given. The default is 1000, which runs quickly. A large value of
                        \code{pginit} makes the function slower.}
  \item{\code{maxiter}}{The maximum number of iterations. The default is 1000.}
  \item{\code{mindiff}}{The minimum difference between two max cosine values to be reached. It defines the convergence criteria:
                         if the difference between the max cosine for the updated parameter values and the previous
                         one is below \code{mindiff}, convergence is reached. The default is 0.000001.}
}
}
\examples{
\dontrun{
  # Example with the bison data set included in the package
  data(bison)
  an <- angular(y.dir ~ y.prec + y.prec2, data = bison)
  print(an)
  summary(an)
  plot(an)
}
}
\references{
L.-P. Rivest, T. Duchesne, A. Nicosia & D. Fortin. A general angular regression model for the analysis of data on animal movement in ecology. Journal of the Royal Statistical Society, series C, to appear.
}
\author{
Sophie Baillargeon, Louis-Paul Rivest, and Aurélien Nicosia
}
