% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_test.r
\name{angular_lrtest}
\alias{angular_lrtest}
\title{Likelihood-ratio test between two homogeneous angular models}
\usage{
angular_lrtest(full, reduced, check = TRUE)
}
\arguments{
\item{full}{An object of class "angular" (the larger model).}

\item{reduced}{An object of class "angular" (the nested model).}

\item{check}{Logical; run basic sanity checks (default TRUE).}
}
\value{
An object of class \code{"htest"} with fields:
\itemize{
  \item \code{statistic} The test statistic \(T_2\).
  \item \code{parameter} The degrees of freedom (difference in number of \eqn{\beta}).
  \item \code{p.value}   The \eqn{\chi^2} p-value.
  \item \code{method}    A short description.
  \item \code{data.name} Model names compared.
  \item \code{details}   List with \code{n}, \code{kappa_full}, \code{Cbar_full}, \code{Cbar_reduced},
                         and \code{LRT_logLik} (= 2*Delta_logLik, if \code{logLik.angular} is available).
}
}
\description{
Likelihood-ratio test between two homogeneous angular models
}
