% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bison.R
\docType{data}
\name{bison}
\alias{bison}
\title{bison}
\format{
Un data frame avec X observations et Y variables:
\describe{
  \item{variable1}{Description de la variable1.}
  \item{variable2}{Description de la variable2.}
}
}
\usage{
data(bison)
}
\description{
Une brève description de ce que contient le jeu de données.
}
\keyword{datasets}
