% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{print.consensus}
\alias{print.consensus}
\alias{coef.consensus}
\alias{residuals.consensus}
\alias{summary.consensus}
\alias{print.summary.consensus}
\alias{plot.consensus}
\alias{AIC.consensus}
\alias{BIC.consensus}
\alias{logLik.consensus}
\title{Methods for Consensus Objects}
\usage{
\method{print}{consensus}(x, ...)

\method{coef}{consensus}(object, ...)

\method{residuals}{consensus}(object, ...)

\method{summary}{consensus}(object, ...)

\method{print}{summary.consensus}(x, ...)

\method{plot}{consensus}(x, ...)

\method{AIC}{consensus}(object, ..., k = 2)

\method{BIC}{consensus}(object, ...)

\method{logLik}{consensus}(object, ...)
}
\arguments{
\item{x}{An object of class "consensus".}

\item{...}{Further arguments passed to plotting functions.}

\item{object}{An object of class "consensus".}
}
\value{
An object of class "summary.consensus", a list containing:
\describe{
  \item{call}{The matched call.}
  \item{MaxLL}{The maximum log-likelihood value.}
  \item{AIC}{The Akaike information criterion.}
  \item{BIC}{The Bayesian information criterion.}
  \item{coefficients}{A matrix with the estimates, robust standard errors, z-values and associated p-values.}
  \item{residuals}{A summary of the residuals (min, 1st Qu., median, 3rd Qu., max).}
  \item{nobs}{The number of observations.}
}
}
\description{
These functions provide methods for printing, extracting coefficients,
summarizing, and retrieving residuals from a consensus model.

This function summarizes an object of class "consensus" by providing key
diagnostic statistics including the model call, maximum log-likelihood, AIC, BIC, a summary of residuals,
the coefficient estimates, and the number of observations.

This function prints a summary of a consensus model.

This function produces a set of diagnostic plots for an object of class "consensus"
in a layout similar to that of \code{plot.lm}. The diagnostics include:
\enumerate{
  \item Residuals vs Fitted values.
  \item Histogram of scaled residuals with a normal density overlay.
  \item Normal Q-Q plot.
  \item A text display of a goodness-of-fit measure: the Spearman correlation between
        the absolute raw residuals and the inverse of the length parameter.
}
}
