% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{decentredPredictorModel}
\alias{decentredPredictorModel}
\title{Decentered Predictor Model Wrapper}
\usage{
decentredPredictorModel(data, response, w, ...)
}
\arguments{
\item{data}{A data frame containing at least the response variable and the explanatory variable.}

\item{response}{A character string representing the name of the response variable (an angular variable).}

\item{w}{A character string representing the name of the explanatory angular variable in \code{data}.}

\item{...}{Additional arguments to be passed to the underlying \code{angular} estimation function.}
}
\value{
An object representing the fitted model as returned by the \code{angular} function.
}
\description{
This function wraps the estimation of the "decentered predictor" model as described in Rivest (1997),
which is based on a single angular explanatory variable. The synthetic model is built using four covariates:
\itemize{
  \item{\code{w}: the original explanatory angular variable,}
  \item{\code{w_plus_pi2}: computed as \eqn{w + \pi/2},}
  \item{\code{zero}: a constant covariate with value 0,}
  \item{\code{pi2}: a constant covariate with value \eqn{\pi/2}.}
}
}
\details{
The model is expressed as:
\deqn{y = w + \left(w + \frac{\pi}{2}\right) + 0 + \left(\frac{\pi}{2}\right),}
which corresponds to the decentered predictor formulation.
}
