% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{decentredPredictorModel}
\alias{decentredPredictorModel}
\title{Decentered Predictor Model Wrapper}
\usage{
decentredPredictorModel(data, response, w, ...)
}
\arguments{
\item{data}{A data frame containing \code{response} and \code{w}.}

\item{response}{A string: name of the angular response variable.}

\item{w}{A string: name of the angular explanatory variable in \code{data}.}

\item{...}{Additional arguments passed to \code{angular()}.}
}
\value{
A \strong{list}: \code{fit}, \code{data_aug}, \code{formula}.
}
\description{
Wraps the "decentered predictor" model (Rivest, 1997) for a single angular
explanatory variable \code{w}. The augmented data include:
\itemize{
  \item \code{<w>_plus_pi2}: \eqn{w + \pi/2};
  \item \code{const0}: 0;
  \item \code{constPi2}: \eqn{\pi/2}.
}
The fitted formula is:
\code{response ~ <w> + <w>_plus_pi2 + const0 + constPi2}.
}
