% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_reference_angle.r
\name{.add_beta1_row}
\alias{.add_beta1_row}
\title{Add the β_ref = 1 row (sd = 0) on top of a 2-column β table}
\usage{
.add_beta1_row(beta_tab, ref_name, put_first = TRUE)
}
\description{
Add the β_ref = 1 row (sd = 0) on top of a 2-column β table
}
