% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_reference_angle.r
\name{.make_angular_formula_with_ref}
\alias{.make_angular_formula_with_ref}
\title{Build an angular formula with \code{ref(angle_ref)} for the chosen reference angle}
\usage{
.make_angular_formula_with_ref(formula, ref_name)
}
\description{
Build an angular formula with \code{ref(angle_ref)} for the chosen reference angle
}
