% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{meanDirectionModel}
\alias{meanDirectionModel}
\title{Mean Direction Model Wrapper}
\usage{
meanDirectionModel(data, response, mu0 = NULL, ...)
}
\arguments{
\item{data}{A data frame containing at least the response variable.}

\item{response}{A character string representing the name of the response variable (an angular variable).}

\item{mu0}{Optional. A numeric vector or value to be used as the baseline covariate.
If \code{NULL} (default), the baseline is set to 0 and \code{tan_mu} is set to pi/2.
Otherwise, \code{tan_mu} is set to \code{mu0 + pi/2}.}

\item{...}{Additional arguments to be passed to the \code{angular} function.}
}
\value{
An object representing the fitted model as returned by the \code{angular} function.
}
\description{
This function wraps the estimation of the "mean direction model" by automatically
creating the required covariates and constructing the appropriate formula.
It generates two covariates:
\itemize{
  \item{\code{mu0}: If \code{mu0} is \code{NULL} (default), this is set to 0. Otherwise, it is set to the provided value (or vector).}
  \item{\code{tan_mu}: If \code{mu0} is \code{NULL}, this is set to \eqn{\pi/2}; otherwise, it is set to \code{mu0 + pi/2}.}
}
}
\details{
The model is then specified as:
\deqn{y = 1 \times mu0 + tan\_mu,}
where the coefficient for \code{mu0} is fixed at 1 (by design in \code{angular}),
and the estimated coefficient for \code{tan_mu} corresponds to \eqn{\tan(\mu)}.
}
