% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{presnellModel}
\alias{presnellModel}
\title{Presnell Model Wrapper}
\usage{
presnellModel(data, response, w, ...)
}
\arguments{
\item{data}{A data frame containing at least the response variable and the continuous covariate w.}

\item{response}{A character string representing the name of the response variable (an angular variable).}

\item{w}{A character string representing the name of the continuous covariate in \code{data}.}

\item{...}{Additional arguments to be passed to the underlying \code{angular} estimation function.}
}
\value{
An object representing the fitted model as returned by the \code{angular} function.
}
\description{
This function wraps the estimation of the "Presnell model" as described in Presnell et al. (1998),
which incorporates one non-circular continuous covariate (w). The synthetic model is specified using
the following form:
\deqn{y = 0 + \frac{\pi}{2} + (0 : w) + \left(\frac{\pi}{2} : w\right),}
where:
\itemize{
  \item The term 0 is represented by a constant covariate,
  \item The term \eqn{\pi/2} is represented by a constant covariate,
  \item 0 : w denotes the interaction between the constant 0 and the continuous covariate w,
  \item \eqn{(\pi/2) : w} denotes the interaction between the constant \eqn{\pi/2} and w.
}
}
\details{
In order for the formula to be interpreted correctly by the underlying \code{angular} function,
this wrapper first creates two variables in \code{data}: \code{zero} (set to 0) and \code{pi2} (set to pi/2).
The model is then specified as:
\deqn{y ~ zero + pi2 + zero:w + pi2:w.}
}
