% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_reference_angle.r
\name{print.pick_reference_angle}
\alias{print.pick_reference_angle}
\title{Print method for \code{pick_reference_angle}}
\usage{
\method{print}{pick_reference_angle}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{pick_reference_angle()}.}

\item{...}{Unused.}
}
\description{
Print method for \code{pick_reference_angle}
}
